* Dont import unnecesarry libs  * Show profile info from comar  * A lot of changes ...

+++ kde4/network-manager/manager/setup.py
@@ -1,4 +1,4 @@
-#!/usr/bin/env python
+#/usr/bin/env python
 #
 # Copyright (C) 2009 TUBITAK/UEKAE
 #
@@ -110,11 +110,9 @@ class Install(install):
         service_file = os.path.join(KDEDIR, "share/kde4/services", DESKTOPFILE)
         shutil.copy(os.path.join("build/",DESKTOPFILE), service_file)
         shutil.move(os.path.join(project_dir,"main.py"), os.path.join(project_dir,PROJECT))
-        try:
+        if not os.path.exists(os.path.join(project_dir,PROJECT+'.py')):
             os.symlink(os.path.join(project_dir,PROJECT), os.path.join(project_dir,PROJECT+'.py'))
             os.symlink(os.path.join(project_dir,PROJECT), os.path.join(KDEDIR,"bin/",PROJECT))
-        except:
-            pass
         os.chmod(os.path.join(project_dir,PROJECT+'.py'),0755)
 
 setup(
